moved image processing out of app.py
